DATA_PUSHER_PROJECT/Data_Pusher/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
appdirs==1.4.3
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.10.0
blinker==1.6.2
CacheControl==0.12.6
caer==2.0.2
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
chatbot==1.5.2b0
click==8.1.3
colorama==0.4.3
comtypes==1.1.10
contextlib2==0.6.0
cryptography==38.0.4
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.6
distro==1.4.0
Django==4.1.1
django-allauth==0.54.0
django-jazzmin==2.6.0
django-oauth-toolkit==2.3.0
django-qrcode==0.3
djangorestframework==3.14.0
dnspython==2.2.1
docutils==0.16
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.96.0
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Limiter==3.3.1
frozenlist==1.3.3
future==0.18.3
geographiclib==2.0
geopy==2.3.0
greenlet==2.0.2
gTTS==2.3.2
h11==0.14.0
html5lib==1.0.1
huepy==1.2.1
idna==2.8
imageio==2.23.0
imageio-ffmpeg==0.4.7
importlib-metadata==6.6.0
importlib-resources==5.12.0
imutils==0.5.4
iniconfig==2.0.0
instabot==0.117.0
ipaddr==2.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jwcrypto==1.5.0
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
limits==3.5.0
lockfile==0.12.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
mock==5.0.2
MouseInfo==0.1.3
moviepy==1.0.3
msgpack==0.6.2
multidict==6.0.4
mypy==0.910
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector-python==8.0.26
mysqlclient==2.0.3
newsapi==0.1.1
nominatim==0.1
nslookup==1.6.1
numpy==1.20.1
oauthlib==3.2.2
opencv-contrib-python==4.5.3.56
openpyxl==3.0.10
ordered-set==4.1.0
packaging==23.1
pandas==1.2.2
passlib==1.7.4
pep517==0.8.2
Pillow==9.2.0
platformdirs==2.6.2
pluggy==1.0.0
proglog==0.1.10
progress==1.5
protobuf==3.18.0
psycopg2==2.9.1
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==1.10.8
pygame==2.4.0
PyGetWindow==0.0.9
Pygments==2.15.1
pyjokes==0.6.0
PyJWT==2.7.0
pymongo==4.3.3
PyMsgBox==1.0.9
pyOpenSSL==22.1.0
pyparsing==2.4.6
PyPDF2==3.0.1
pyperclip==1.8.2
pypiwin32==223
PyQRCode==1.2.1
PyRect==0.1.4
PyScreeze==0.1.28
PySocks==1.7.1
pytest==7.3.1
python-dateutil==2.8.1
python3-openid==3.2.0
pythonping==1.1.4
pytoml==0.1.21
pyttsx3==2.90
pytweening==1.0.4
pytz==2020.5
pywhatkit==5.1
pywin32==300
PyYAML==6.0
qrcode==7.3.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.23.1
retrying==1.3.3
rich==13.4.1
schedule==1.2.0
scikit-learn==1.2.2
scipy==1.7.1
six==1.14.0
sniffio==1.3.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
SQLAlchemy==2.0.15
sqlparse==0.4.1
starlette==0.27.0
telepot==12.7
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
types-PyYAML==6.0.12.9
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.25.8
uvicorn==0.22.0
virtualenv==20.16.5
webencodings==0.5.1
Werkzeug==2.3.5
whitenoise==5.2.0
wikipedia==1.4.0
wrapt==1.15.0
xlrd==2.0.1
XlsxWriter==3.0.3
xlutils==2.0.0
xlwt==1.3.0
yarl==1.9.2
zipp==3.15.0
